vercel_deploy_only/requirements.txt
python-telegram-bot==20.7
flask==2.3.3 